+++ public_transport_watcher/extractor/configuration/extraction_config.py
@@ -10,4 +10,7 @@ EXTRACTION_CONFIG = {
     "stations": {
         "batch_size": 100,
     },
+    "addresses": {
+        "batch_size": 1000,
+    },
 }