+++ public_transport_watcher/extractor/insert/__init__.py
@@ -1,7 +1,9 @@
+from .addresses import insert_addresses_informations
 from .navigo import insert_navigo_data
 from .stations import insert_stations_informations
 
 __all__ = [
+    "insert_addresses_informations",
     "insert_navigo_data",
     "insert_stations_informations",
 ]