+++ public_transport_watcher/extractor/sql/get_existing_stations.sql
@@ -1,0 +1,4 @@
+SELECT
+    DISTINCT id, name, latitude, longitude
+FROM
+    transport.station;