+++ CHANGELOG.md
@@ -10,9 +10,11 @@ MR #4
 
 ### Added
 - Stations extraction
+- Utility function to execute a `SELECT` query and return the result
 
 ### Changed
 - The datalake file retriever function to accept only one folder in the path
+- Extraction configuration setup
 
 ## [0.2.0] - 2025-03-29
 MR #2